Split main.jsx
- Split main.jsx into multiple modules
- Change tsconfig.json such that React does not need to be imported in
  every file that contains JSX
